Updated the definition of Omnichannel application
Key aspect of omnichannel is enabling continuity of interactions across multiple channels for a user. This is missed out in the definition. Also mobile and web apps are not examples of omnichannel applications, unless they are designed to provide the omnichannel experience. Merely allowing users to interact over multiple channels (without ensuring continuity) will be multichannel and not omnichannel .