Updated maintenance window duration
The content in the table for the "Progression phase tasks" needs to be corrected. 30-day window should be changed to 21-day window. I found the two instances in the same row of the table, and made the change. I am proposing this as a pull request to ensure that I did not change anything else I was not supposed to.